requirements.txt
streamlit==1.33.0
tensorflow==2.13.0
opencv-python-headless==4.8.0.76
pillow==9.5.0
numpy==1.23.5